+++ gems/aws-sdk-s3/spec/object_summary_spec.rb
@@ -53,6 +53,12 @@ module Aws
         object_summary.upload_file('source', opt_name: 'opt-value')
       end
 
+      it 'delegates #upload_chunks to #object' do
+        block = proc { :foo }
+        expect(object).to receive(:upload_chunks).with(opt_name: 'opt-value', &block)
+        object_summary.upload_chunks(opt_name: 'opt-value', &block)
+      end
+
       it 'delegates #download_file to #object' do
         expect(object).to receive(:download_file).with('destination', opt_name: 'opt-value')
         object_summary.download_file('destination', opt_name: 'opt-value')